requirements.txt
# Python 3.9+
# pip install -r requirements.txt
# playwright install chromium
playwright>=1.40,<2
pandas>=2.0,<3
openpyxl>=3.1,<4
matplotlib>=3.7,<4
python-dotenv>=1.0,<2
beautifulsoup4>=4.12,<5
